+++ src/main/java/nl/tudelft/ti2806/pl1/mutation/InDelMutation.java
@@ -1,4 +1,4 @@
-package mutation;
+package nl.tudelft.ti2806.pl1.mutation;
 
 /**
  * @author Maarten, Justin
@@ -7,7 +7,10 @@ package mutation;
 public class InDelMutation extends Mutation {
 
 	/**
-	 * 
+	 * @param pre
+	 *            The ID of the node before the mutation.
+	 * @param post
+	 *            The ID of the node after the mutation.
 	 */
 	public InDelMutation(final int pre, final int post) {
 		super(pre, post);
