+++ src/main/java/nl/tudelft/ti2806/pl1/mutation/Mutation.java
@@ -1,4 +1,4 @@
-package mutation;
+package nl.tudelft.ti2806.pl1.mutation;
 
 /**
  * @author Maarten, Justin
@@ -7,10 +7,17 @@ package mutation;
  */
 public abstract class Mutation {
 
+	/**
+	 * The IDs of the nodes before and after the mutation.
+	 */
 	private int preNode, postNode;
 
 	/**
 	 * 
+	 * @param pre
+	 *            The ID of the node before the mutation.
+	 * @param post
+	 *            The ID of the node after the mutation.
 	 */
 	public Mutation(final int pre, final int post) {
 		this.preNode = pre;
