+++ src/main/java/nl/tudelft/ti2806/pl1/mutation/package-info.java
@@ -2,4 +2,4 @@
  * @author Maarten, Justin
  * @since 2-6-2015
  */
-package mutation;
+package nl.tudelft.ti2806.pl1.mutation;
