+++ src/main/java/nl/tudelft/ti2806/pl1/zoomlevels/PointGraphConverter.java
@@ -52,7 +52,7 @@ public final class PointGraphConverter {
 			while (leaving.hasNext()) {
 				Edge out = leaving.next();
 				Node outnode = out.getNode1();
-				String content; // outnode.getAttribute("content").toString();
+				String content;
 				if (outnode.getAttribute("content") instanceof HashMap<?, ?>) {
 					content = outnode.getAttribute("content").toString();
 				} else {
